+++ internal/services/azapi_resource_action_data_source_test.go
@@ -5,7 +5,6 @@ import (
 	"testing"
 
 	"github.com/Azure/terraform-provider-azapi/internal/acceptance"
-	"github.com/Azure/terraform-provider-azapi/internal/acceptance/check"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
 )
 
@@ -18,9 +17,7 @@ func TestAccActionDataSource_basic(t *testing.T) {
 	data.DataSourceTest(t, []resource.TestStep{
 		{
 			Config: r.basic(data),
-			Check: resource.ComposeTestCheckFunc(
-				check.That(data.ResourceName).Key("output").IsJson(),
-			),
+			Check:  resource.ComposeTestCheckFunc(),
 		},
 	})
 }
@@ -32,9 +29,7 @@ func TestAccActionDataSource_providerPermissions(t *testing.T) {
 	data.DataSourceTest(t, []resource.TestStep{
 		{
 			Config: r.providerPermissions(),
-			Check: resource.ComposeTestCheckFunc(
-				check.That(data.ResourceName).Key("output").IsJson(),
-			),
+			Check:  resource.ComposeTestCheckFunc(),
 		},
 	})
 }
@@ -46,37 +41,7 @@ func TestAccActionDataSource_providerAction(t *testing.T) {
 	data.DataSourceTest(t, []resource.TestStep{
 		{
 			Config: r.providerAction(),
-			Check: resource.ComposeTestCheckFunc(
-				check.That(data.ResourceName).Key("output").IsJson(),
-			),
-		},
-	})
-}
-
-func TestAccActionDataSource_dynamicSchema(t *testing.T) {
-	data := acceptance.BuildTestData(t, "data.azapi_resource_action", "test")
-	r := ActionDataSource{}
-
-	data.DataSourceTest(t, []resource.TestStep{
-		{
-			Config: r.dynamicSchema(),
-			Check: resource.ComposeTestCheckFunc(
-				check.That(data.ResourceName).Key("output").IsJson(),
-			),
-		},
-	})
-}
-
-func TestAccActionDataSource_dynamicSchemaHclOutput(t *testing.T) {
-	data := acceptance.BuildTestData(t, "data.azapi_resource_action", "test")
-	r := ActionDataSource{}
-
-	data.DataSourceTest(t, []resource.TestStep{
-		{
-			Config: r.dynamicSchemaHclOutput(),
-			Check: resource.ComposeTestCheckFunc(
-				check.That(data.ResourceName).Key("output.count").Exists(),
-			),
+			Check:  resource.ComposeTestCheckFunc(),
 		},
 	})
 }
@@ -113,38 +78,6 @@ data "azapi_resource_action" "test" {
 
 func (r ActionDataSource) providerAction() string {
 	return `
-data "azapi_resource_action" "test" {
-  type        = "Microsoft.ResourceGraph@2020-04-01-preview"
-  resource_id = "/providers/Microsoft.ResourceGraph"
-  action      = "resources"
-  body = jsonencode({
-    query = "resources| limit 1"
-  })
-  response_export_values = ["*"]
-}
-`
-}
-
-func (r ActionDataSource) dynamicSchema() string {
-	return `
-data "azapi_resource_action" "test" {
-  type        = "Microsoft.ResourceGraph@2020-04-01-preview"
-  resource_id = "/providers/Microsoft.ResourceGraph"
-  action      = "resources"
-  body = {
-    query = "resources| limit 1"
-  }
-  response_export_values = ["*"]
-}
-`
-}
-
-func (r ActionDataSource) dynamicSchemaHclOutput() string {
-	return `
-provider "azapi" {
-  enable_hcl_output_for_data_source = true
-}
-
 data "azapi_resource_action" "test" {
   type        = "Microsoft.ResourceGraph@2020-04-01-preview"
   resource_id = "/providers/Microsoft.ResourceGraph"