+++ internal/services/azapi_resource_action_resource_test.go
@@ -65,19 +65,7 @@ func TestAccActionResource_providerAction(t *testing.T) {
 
 	data.DataSourceTest(t, []resource.TestStep{
 		{
-			Config: r.providerAction(),
-			Check:  resource.ComposeTestCheckFunc(),
-		},
-	})
-}
-
-func TestAccActionResource_dynamicSchema(t *testing.T) {
-	data := acceptance.BuildTestData(t, "azapi_resource_action", "test")
-	r := ActionResource{}
-
-	data.DataSourceTest(t, []resource.TestStep{
-		{
-			Config: r.dynamicSchema(),
+			Config: r.providerAction(data),
 			Check:  resource.ComposeTestCheckFunc(),
 		},
 	})
@@ -122,9 +110,9 @@ resource "azapi_resource_action" "test" {
   type        = "Microsoft.Automation/automationAccounts@2021-06-22"
   resource_id = azapi_resource.test.id
   action      = "agentRegistrationInformation/regenerateKey"
-  body = jsonencode({
+  body = {
     keyName = "primary"
-  })
+  }
   depends_on = [
     data.azapi_resource_action.list
   ]
@@ -148,9 +136,9 @@ resource "azapi_resource_action" "test" {
   resource_id = azapi_resource.test.id
   when        = "destroy"
   action      = "agentRegistrationInformation/regenerateKey"
-  body = jsonencode({
+  body = {
     keyName = "primary"
-  })
+  }
   depends_on = [
     data.azapi_resource_action.list
   ]
@@ -176,28 +164,8 @@ resource "azapi_resource_action" "test" {
 `
 }
 
-func (r ActionResource) providerAction() string {
-	return `
-provider "azurerm" {
-  features {}
-}
-
-data "azurerm_client_config" "current" {}
-
-resource "azapi_resource_action" "test" {
-  type        = "Microsoft.Cache@2023-04-01"
-  resource_id = "/subscriptions/${data.azurerm_client_config.current.subscription_id}/providers/Microsoft.Cache"
-  action      = "CheckNameAvailability"
-  body = jsonencode({
-    type = "Microsoft.Cache/Redis"
-    name = "cacheName"
-  })
-}
-`
-}
-
-func (r ActionResource) dynamicSchema() string {
-	return `
+func (r ActionResource) providerAction(data acceptance.TestData) string {
+	return fmt.Sprintf(`
 provider "azurerm" {
   features {}
 }
@@ -210,10 +178,10 @@ resource "azapi_resource_action" "test" {
   action      = "CheckNameAvailability"
   body = {
     type = "Microsoft.Cache/Redis"
-    name = "cacheName"
+    name = "%s"
   }
 }
-`
+`, data.RandomString)
 }
 
 func (r ActionResource) nonstandardLRO(data acceptance.TestData) string {
@@ -275,12 +243,12 @@ resource "azapi_resource_action" "test" {
   resource_id            = data.azapi_resource_id.functionKey.id
   method                 = "PUT"
   response_export_values = ["*"]
-  body = jsonencode({
+  body = {
     properties = {
       name  = "test_key"
       value = "test_value"
     }
-  })
+  }
 }
 `, data.LocationPrimary, data.RandomString)
 }
@@ -300,9 +268,9 @@ resource "azapi_resource_action" "test" {
   type        = "Microsoft.Automation/automationAccounts@2021-06-22"
   resource_id = azapi_resource.test.id
   action      = "agentRegistrationInformation/regenerateKey"
-  body = jsonencode({
+  body = {
     keyName = "primary"
-  })
+  }
   depends_on = [
     data.azapi_resource_action.list
   ]