+++ internal/services/azapi_resource_data_source_test.go
@@ -28,7 +28,7 @@ func TestAccGenericDataSource_basic(t *testing.T) {
 				check.That(data.ResourceName).Key("identity.0.tenant_id").Exists(),
 				check.That(data.ResourceName).Key("location").HasValue(location.Normalize(data.LocationPrimary)),
 				check.That(data.ResourceName).Key("tags.%").HasValue("1"),
-				check.That(data.ResourceName).Key("output").IsJson(),
+				check.That(data.ResourceName).Key("output.properties.%").Exists(),
 			),
 		},
 	})
@@ -49,7 +49,7 @@ func TestAccGenericDataSource_withResourceId(t *testing.T) {
 				check.That(data.ResourceName).Key("identity.0.tenant_id").Exists(),
 				check.That(data.ResourceName).Key("location").HasValue(location.Normalize(data.LocationPrimary)),
 				check.That(data.ResourceName).Key("tags.%").HasValue("1"),
-				check.That(data.ResourceName).Key("output").IsJson(),
+				check.That(data.ResourceName).Key("output.properties.%").Exists(),
 			),
 		},
 	})
@@ -65,21 +65,6 @@ func TestAccGenericDataSource_defaultParentId(t *testing.T) {
 			Config: r.defaultParentId(data),
 			Check: resource.ComposeTestCheckFunc(
 				check.That(data.ResourceName).Key("parent_id").HasValue(fmt.Sprintf("/subscriptions/%s", subscriptionId)),
-				check.That(data.ResourceName).Key("output").IsJson(),
-			),
-		},
-	})
-}
-
-func TestAccGenericDataSource_hclOutput(t *testing.T) {
-	data := acceptance.BuildTestData(t, "data.azapi_resource", "test")
-	r := GenericDataSource{}
-
-	data.DataSourceTest(t, []resource.TestStep{
-		{
-			Config: r.hclOutput(data),
-			Check: resource.ComposeTestCheckFunc(
-				check.That(data.ResourceName).Key("output.properties.%").Exists(),
 			),
 		},
 	})
@@ -133,22 +118,6 @@ data "azapi_resource" "test" {
 func (r GenericDataSource) withResourceId(data acceptance.TestData) string {
 	return fmt.Sprintf(`
 %s
-
-data "azapi_resource" "test" {
-  type                   = azapi_resource.test.type
-  resource_id            = azapi_resource.test.id
-  response_export_values = ["*"]
-}
-`, GenericResource{}.complete(data))
-}
-
-func (r GenericDataSource) hclOutput(data acceptance.TestData) string {
-	return fmt.Sprintf(`
-%s
-
-provider "azapi" {
-  enable_hcl_output_for_data_source = true
-}
 
 data "azapi_resource" "test" {
   type                   = azapi_resource.test.type