+++ internal/services/azapi_resource_list_data_source_test.go
@@ -18,20 +18,6 @@ func TestAccListDataSource_basic(t *testing.T) {
 		{
 			Config: r.basic(),
 			Check: resource.ComposeTestCheckFunc(
-				check.That(data.ResourceName).Key("output").IsJson(),
-			),
-		},
-	})
-}
-
-func TestAccListDataSource_hclOutput(t *testing.T) {
-	data := acceptance.BuildTestData(t, "data.azapi_resource_list", "test")
-	r := ListDataSource{}
-
-	data.DataSourceTest(t, []resource.TestStep{
-		{
-			Config: r.hclOutput(),
-			Check: resource.ComposeTestCheckFunc(
 				check.That(data.ResourceName).Key("output.value.#").Exists(),
 			),
 		},
@@ -54,26 +40,6 @@ func (r ListDataSource) basic() string {
 	return `
 provider "azurerm" {
   features {}
-}
-
-data "azurerm_client_config" "current" {}
-
-data "azapi_resource_list" "test" {
-  type                   = "Microsoft.Resources/resourceGroups@2024-03-01"
-  parent_id              = "/subscriptions/${data.azurerm_client_config.current.subscription_id}"
-  response_export_values = ["*"]
-}
-`
-}
-
-func (r ListDataSource) hclOutput() string {
-	return `
-provider "azurerm" {
-  features {}
-}
-
-provider "azapi" {
-  enable_hcl_output_for_data_source = true
 }
 
 data "azurerm_client_config" "current" {}