+++ quickstarts/Microsoft.Web/list_and_update_appsettings/main.tf
@@ -59,7 +59,7 @@ data "azapi_resource" "appsettings" {
 
 output "o1" {
   // appsettings can't be fetched with azapi_resource data source directly
-  value = jsondecode(data.azapi_resource.appsettings.output)
+  value = data.azapi_resource.appsettings.output
 }
 
 data "azapi_resource_action" "list" {
@@ -72,26 +72,26 @@ data "azapi_resource_action" "list" {
 
 output "o2" {
   // appsettings can only be fetched with list action
-  value = jsondecode(data.azapi_resource_action.list.output)
+  value = data.azapi_resource_action.list.output
 }
 
 resource "azapi_resource_action" "update" {
   type        = "Microsoft.Web/sites/config@2022-03-01"
   resource_id = data.azapi_resource.appsettings.id
   method      = "PUT"
-  body = jsonencode({
+  body = {
     name = "appsettings"
     // use merge function to combine new settings with existing ones
     properties = merge(
-      jsondecode(data.azapi_resource_action.list.output).properties,
+      data.azapi_resource_action.list.output.properties,
       {
         WEBSITES_ENABLE_APP_SERVICE_STORAGE = "false"
       }
     )
-  })
+  }
   response_export_values = ["*"]
 }
 
 output "o3" {
-  value = jsondecode(azapi_resource_action.update.output)
+  value = azapi_resource_action.update.output
 }