+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## v2.0.0-beta
+BREAKING CHANGES:
+- Provider field `default_naming_prefix` and `default_naming_suffix` are removed. Please specify the naming prefix and suffix in the resource's `name` field instead.
+- Provider field `enable_hcl_output_for_data_source` is removed. The `output` field in the data source is always in HCL format.
+- The `azapi_resource`'s `removing_special_chars` field is removed. Please specify the `name` field and remove the special characters in the `name` field instead.
+- The `ignore_body_changes` field is removed. Please use the `lifecyle.ignore_changes` to ignore some properties when comparing the resource with its current state.
+- The `body` field now only accepts an HCL object. Please remove the `jsondecode` function when using the `body` field.
+- The `output` field now only exports an HCL object. Please remove the `jsondecode` function when using the `output` field.
+
 ## v1.15.0
 
 ENHANCEMENTS:
@@ -31,8 +40,8 @@ BUG FIXES:
 ENHANCEMENTS:
 - `azapi` provider: Support `enable_hcl_output_for_data_source` field, which is used to enable the HCL output for the data source, the default value is `false`.
   This could resolve the following breaking changes in the previous release:
-  - `azapi_resource` data source: The `output` field changes from JSON string to HCL object. Users can use access the fields in the output as an HCL object. Please remove the `jsondecode` function when using the `output` field.
-  - `azapi_resource_list` data source: The `output` field changes from JSON string to HCL object. Users can use access the fields in the output as an HCL object. Please remove the `jsondecode` function when using the `output` field.
+- `azapi_resource` data source: The `output` field changes from JSON string to HCL object. Users can use access the fields in the output as an HCL object. Please remove the `jsondecode` function when using the `output` field.
+- `azapi_resource_list` data source: The `output` field changes from JSON string to HCL object. Users can use access the fields in the output as an HCL object. Please remove the `jsondecode` function when using the `output` field.
 
 BUG FIXES:
 - Fix a bug when upgrading from previous provider `azapi_resource` resource will set `tags` for resources that don't have `tags` in the configuration.